+++ crates/motsu/src/lib.rs
@@ -21,11 +21,30 @@
 //!     #[motsu::test]
 //!     fn reads_balance(
 //!         contract: Contract<Erc20>,
-//!         alice: Account,
+//!         alice: Address,
 //!     ) {
 //!         // Access storage.
 //!         let balance = contract.sender(alice).balance_of(Address::ZERO);
 //!         assert_eq!(balance, U256::ZERO);
+//!     }
+//! }
+//! ```
+//!
+//! If you need to instantiate an accound that contains a signer and a private
+//! key, you can use [`crate::prelude::Account`] instead of
+//! [`stylus_sdk::alloy_primitives::Address`]:
+//!
+//! ```rust,ignore
+//! #[cfg(test)]
+//! mod tests {
+//!     use motsu::prelude::*;
+//!     use alloy_signer::SignerSync;
+//!
+//!     #[motsu::test]
+//!     fn signs_message(alice: Account) {
+//!         let msg = "message".as_bytes();
+//!         let signer = alice.signer();
+//!         assert!(signer.sign_message_sync(msg).is_ok());
 //!     }
 //! }
 //! ```
@@ -45,7 +64,7 @@
 //! #[motsu::test]
 //! fn test_with_custom_chain_id(
 //!     contract: Contract<MyContract>,
-//!     alice: Account,
+//!     alice: Address,
 //! ) {
 //!     // Default chain ID is 42161 (Arbitrum One)
 //!
@@ -74,7 +93,7 @@
 //!  use stylus_sdk::alloy_primitives::{Address, U256, ruint::uint};
 //!
 //!  #[motsu::test]
-//!  fn pay_three_proxies(proxy: Contract<Proxy>, alice: Account) {
+//!  fn pay_three_proxies(proxy: Contract<Proxy>, alice: Address) {
 //!     let one = uint!(1_U256);
 //!     let ten = uint!(10_U256);
 //!
@@ -96,9 +115,9 @@
 //! ### External Calls
 //!
 //! Multiple external calls are supported in Motsu.
-//! Assuming `Proxy` is a contract that exposes `#[public]` function
-//! `Proxy::call_proxy`, where it adds `one` to the passed argument and calls
-//! next `Proxy` contract at the address provided during initialization.
+//! Assuming `Proxy` is a contract that exposes [`stylus_sdk::prelude::public`]
+//! function `Proxy::call_proxy`, where it adds `one` to the passed argument and
+//! calls next `Proxy` contract at the address provided during initialization.
 //! The following test case can emulate a call chain of three `Proxy` contracts:
 //!
 //! ```rust,ignore
@@ -110,7 +129,7 @@
 //!     proxy1: Contract<Proxy>,
 //!     proxy2: Contract<Proxy>,
 //!     proxy3: Contract<Proxy>,
-//!     alice: Account,
+//!     alice: Address,
 //!  ) {
 //!     let one = uint!(1_U256);
 //!     let ten = uint!(10_U256);
@@ -380,7 +399,7 @@ mod ping_pong_tests {
     fn external_call(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         let value = TEN;
         let ponged_value =
@@ -395,7 +414,7 @@ mod ping_pong_tests {
     fn external_call_error(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         let value = MAGIC_ERROR_VALUE;
         let err =
@@ -411,7 +430,7 @@ mod ping_pong_tests {
     fn external_call_panic(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         let value = MAGIC_ERROR_VALUE;
         ping.sender(alice).ping(pong.address(), value).motsu_unwrap();
@@ -421,7 +440,7 @@ mod ping_pong_tests {
     fn external_static_call(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         let can_ping =
             ping.sender(alice).can_ping(pong.address()).motsu_unwrap();
@@ -432,14 +451,14 @@ mod ping_pong_tests {
     fn msg_sender(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         assert_eq!(ping.sender(alice).pinged_from.get(), Address::ZERO);
         assert_eq!(pong.sender(alice).ponged_from.get(), Address::ZERO);
 
         ping.sender(alice).ping(pong.address(), TEN).motsu_unwrap();
 
-        assert_eq!(ping.sender(alice).pinged_from.get(), alice.address());
+        assert_eq!(ping.sender(alice).pinged_from.get(), alice);
         assert_eq!(pong.sender(alice).ponged_from.get(), ping.address());
     }
 
@@ -447,7 +466,7 @@ mod ping_pong_tests {
     fn has_code(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         assert!(ping.sender(alice).has_pong(pong.address()));
     }
@@ -456,7 +475,7 @@ mod ping_pong_tests {
     fn contract_address(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         assert_eq!(ping.sender(alice).contract_address.get(), Address::ZERO);
         assert_eq!(pong.sender(alice).contract_address.get(), Address::ZERO);
@@ -471,7 +490,7 @@ mod ping_pong_tests {
     fn deref_invalidated_storage_cache(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         // This test is different from `contract_address` test, since it checks
         // that `StorageType` contract correctly resets.
@@ -508,13 +527,13 @@ mod ping_pong_tests {
     // chain, this is still an allowed "feature" of motsu, so we "document" the
     // behavior with this unit test.
     #[motsu::test]
-    fn storage_cleanup(alice: Account, addr: Address) {
+    fn storage_cleanup(alice: Address, addr: Address) {
         // First contract
         let pong1 = Contract::<PongContract>::new_at(addr);
 
         pong1.sender(alice).pong(U256::ZERO).expect("should pong");
 
-        assert_eq!(alice.address(), pong1.sender(alice).ponged_from.get());
+        assert_eq!(alice, pong1.sender(alice).ponged_from.get());
         assert_eq!(ONE, pong1.sender(alice).pongs_count.get());
         assert_eq!(addr, pong1.sender(alice).contract_address.get());
 
@@ -534,20 +553,18 @@ mod ping_pong_tests {
     fn emits_event(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         ping.sender(alice).ping(pong.address(), TEN).motsu_unwrap();
 
         // Assert emitted events.
-        ping.assert_emitted(&Pinged { from: alice.address(), value: TEN });
+        ping.assert_emitted(&Pinged { from: alice, value: TEN });
         pong.assert_emitted(&Ponged { from: ping.address(), value: TEN });
 
         // Assert not emitted events
         assert!(!ping.emitted(&Pinged { from: ping.address(), value: TEN }));
-        assert!(!pong.emitted(&Ponged { from: alice.address(), value: TEN }));
-        assert!(
-            !ping.emitted(&Pinged { from: alice.address(), value: TEN + ONE })
-        );
+        assert!(!pong.emitted(&Ponged { from: alice, value: TEN }));
+        assert!(!ping.emitted(&Pinged { from: alice, value: TEN + ONE }));
         assert!(
             !pong.emitted(&Ponged { from: ping.address(), value: TEN + ONE })
         );
@@ -560,7 +577,7 @@ mod ping_pong_tests {
     fn asserted_emitted_event(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         ping.sender(alice).ping(pong.address(), TEN).motsu_unwrap();
 
@@ -576,17 +593,17 @@ mod ping_pong_tests {
     fn assert_reverts_emitted_event(
         ping: Contract<PingContract>,
         pong: Contract<PongContract>,
-        alice: Account,
+        alice: Address,
     ) {
         let value = MAGIC_ERROR_VALUE;
         ping.sender(alice).ping(pong.address(), value).motsu_unwrap_err();
 
         // Both events should not be emitted after revert.
-        assert!(!ping.emitted(&Pinged { from: alice.address(), value }));
+        assert!(!ping.emitted(&Pinged { from: alice, value }));
         assert!(!pong.emitted(&Ponged { from: ping.address(), value }));
 
         // Check panic assertion.
-        ping.assert_emitted(&Pinged { from: alice.address(), value });
+        ping.assert_emitted(&Pinged { from: alice, value });
     }
 }
 
@@ -836,7 +853,7 @@ mod proxies_tests {
         proxy1: Contract<Proxy>,
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of three proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3.
@@ -856,7 +873,7 @@ mod proxies_tests {
         proxy1: Contract<Proxy>,
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of three proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3 -> proxy1 -> ..
@@ -876,7 +893,7 @@ mod proxies_tests {
         proxy1: Contract<Proxy>,
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of three proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3.
@@ -905,7 +922,7 @@ mod proxies_tests {
         proxy1: Contract<Proxy>,
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of three proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3.
@@ -935,7 +952,7 @@ mod proxies_tests {
         proxy1: Contract<Proxy>,
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of three proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3.
@@ -968,7 +985,7 @@ mod proxies_tests {
             let proxy1 = Contract::<Proxy>::new();
             let proxy2 = Contract::<Proxy>::new();
             let proxy3 = Contract::<Proxy>::new();
-            let alice = Account::random();
+            let alice = Address::random();
 
             // Set up a chain of three proxies.
             // With the given call chain: proxy1 -> proxy2 -> proxy3.
@@ -990,7 +1007,7 @@ mod proxies_tests {
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
         proxy4: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of four proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3 -> proxy4.
@@ -1036,7 +1053,7 @@ mod proxies_tests {
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
         proxy4: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of four proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3 -> proxy4.
@@ -1087,7 +1104,7 @@ mod proxies_tests {
         proxy2: Contract<Proxy>,
         proxy3: Contract<Proxy>,
         proxy4: Contract<Proxy>,
-        alice: Account,
+        alice: Address,
     ) {
         // Set up a chain of four proxies.
         // With the given call chain: proxy1 -> proxy2 -> proxy3 -> proxy4.
@@ -1130,10 +1147,10 @@ mod proxies_tests {
 
 #[cfg(test)]
 mod vm_tests {
-    use stylus_sdk::{block, prelude::*};
+    use stylus_sdk::{alloy_primitives::Address, block, prelude::*};
 
     use crate as motsu;
-    use crate::prelude::*;
+    use crate::{context::DEFAULT_CHAIN_ID, prelude::*};
 
     const ETHEREUM_SEPOLIA_CHAIN_ID: u64 = 11155111;
     const CUSTOM_CHAIN_ID: u64 = 12345678987654321;
@@ -1151,7 +1168,7 @@ mod vm_tests {
     unsafe impl TopLevelStorage for ChainChecker {}
 
     #[motsu::test]
-    fn chain_id(contract: Contract<ChainChecker>, alice: Account) {
+    fn chain_id(contract: Contract<ChainChecker>, alice: Address) {
         // Default chain ID is Arbitrum One
         let chain_id = contract.sender(alice).get_chain_id();
         assert_eq!(chain_id, DEFAULT_CHAIN_ID);