+++ crates/motsu/src/prelude.rs
@@ -1,8 +1,7 @@
 //! Common imports for `motsu` tests.
 pub use crate::{
     context::{
-        Account, Contract, ContractCall, FromTag, Funding, VMContext,
-        DEFAULT_CHAIN_ID,
+        Account, Balance, Contract, ContractCall, FromTag, Funding, VMContext,
     },
     revert::ResultExt,
 };