+++ CHANGELOG.md
@@ -9,10 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added ability to set the chain ID in tests #67
+- `Account` can be created from string and byte seeds, and can return a signer. #70
+- Implemented `Balance` trait for `Account` and `Contract<T>`. #70
+- Added ability to set the chain ID in tests. #67
 
 ### Changed (Breaking)
 
+- `Account` can no longer be created at a predetermined address. #70
+- `Funding::balance` is removed. #70
 - Bump Stylus SDK to v0.8.3-rc.1 #75
 
 ## [0.6.0] - 2025-03-11